requirements.txt
flask
flask-cors
azure-ai-formrecognizer
python-dateutil
